feat(plugin-image): expose  prop in Editor config
