+++ apps/web/src/serlo-editor-integration/image-with-serlo-config.ts
@@ -1,5 +1,9 @@
-import { LoadedFile, UploadValidator } from '@editor/plugin'
+import { LoadedFile } from '@editor/plugin'
 import { createImagePlugin } from '@editor/plugins/image'
+import {
+  type FileError,
+  validateFile,
+} from '@editor/plugins/image/utils/validate-file'
 import { gql } from 'graphql-request'
 
 import { createAuthAwareGraphqlFetch } from '@/api/graphql-fetch'
@@ -8,8 +12,6 @@ import { fetchAndPersistAuthSession } from '@/auth/cookie/fetch-and-persist-auth
 import { MediaType, MediaUploadQuery } from '@/fetcher/graphql-types/operations'
 import { showToastNotice } from '@/helper/show-toast-notice'
 
-const maxFileSize = 2 * 1024 * 1024
-const allowedExtensions = ['gif', 'jpg', 'jpeg', 'png', 'svg', 'webp']
 const supportedMimeTypes = [
   'image/gif',
   'image/jpeg',
@@ -28,41 +30,12 @@ const mimeTypesToMediaType: Record<SupportedMimeType, MediaType> = {
   'image/webp': MediaType.ImageWebp,
 }
 
-enum FileErrorCode {
-  TOO_MANY_FILES,
-  NO_FILE_SELECTED,
-  BAD_EXTENSION,
-  FILE_TOO_BIG,
-  UPLOAD_FAILED,
-}
-
-export interface FileError {
-  errorCode: FileErrorCode
-  message: string
-}
-
-const validateFile: UploadValidator<FileError[]> = (file) => {
-  let uploadErrors: FileErrorCode[] = []
-
-  if (!file) {
-    uploadErrors = [...uploadErrors, FileErrorCode.NO_FILE_SELECTED]
-  } else if (!matchesAllowedExtensions(file.name)) {
-    uploadErrors = [...uploadErrors, FileErrorCode.BAD_EXTENSION]
-  } else if (file.size > maxFileSize) {
-    uploadErrors = [...uploadErrors, FileErrorCode.FILE_TOO_BIG]
-  } else {
-    return { valid: true }
-  }
-
-  return { valid: false, errors: handleErrors(uploadErrors) }
-}
-
 export const imagePlugin = createImagePlugin({
   upload: createUploadImageHandler(),
   validate: validateFile,
 })
 
-function createUploadImageHandler() {
+export function createUploadImageHandler() {
   const readFile = createReadFile()
   return async function uploadImageHandler(file: File): Promise<string> {
     const validation = validateFile(file)
@@ -126,35 +99,8 @@ export function createReadFile() {
   }
 }
 
-function matchesAllowedExtensions(fileName: string) {
-  const extension = fileName.toLowerCase().slice(fileName.lastIndexOf('.') + 1)
-  return allowedExtensions.includes(extension)
-}
-
-function handleErrors(errors: FileErrorCode[]): FileError[] {
-  return errors.map((error) => ({
-    errorCode: error,
-    message: errorCodeToMessage(error),
-  }))
-}
-
 function onError(errors: FileError[]): void {
   showToastNotice(errors.map((error) => error.message).join('\n'), 'warning')
-}
-
-function errorCodeToMessage(error: FileErrorCode) {
-  switch (error) {
-    case FileErrorCode.TOO_MANY_FILES:
-      return 'You can only upload one file'
-    case FileErrorCode.NO_FILE_SELECTED:
-      return 'No file selected'
-    case FileErrorCode.BAD_EXTENSION:
-      return 'Not an accepted file type'
-    case FileErrorCode.FILE_TOO_BIG:
-      return 'Filesize is too big'
-    case FileErrorCode.UPLOAD_FAILED:
-      return 'Error while uploading'
-  }
 }
 
 const uploadUrlQuery = gql`
