+++ packages/editor/src/package/config.ts
@@ -1,6 +1,8 @@
 import type { PluginWithData } from '@editor/plugin/helpers/editor-plugins'
 import type { PluginStaticRenderer } from '@editor/plugin/helpers/editor-renderer'
 import type { BoxConfig } from '@editor/plugins/box'
+import { ImageConfig } from '@editor/plugins/image'
+import { validateFile } from '@editor/plugins/image/utils/validate-file'
 import {
   defaultConfig as defaultMultimediaConfig,
   type MultimediaConfig,
@@ -12,6 +14,10 @@ import type { SupportedLanguage } from '@editor/types/language-data'
 
 export interface PluginsConfig {
   box?: BoxConfig
+  image?: {
+    validate?: ImageConfig['validate']
+    upload: ImageConfig['upload']
+  }
   multimedia?: MultimediaConfig
   spoiler?: SpoilerConfig
   table?: SerloTableConfig
@@ -25,9 +31,16 @@ export interface PluginsConfig {
 // and will not be supported in the future
 export type CustomPlugin = PluginWithData & PluginStaticRenderer
 
-const defaultPluginsConfig: Required<PluginsConfig> = {
+type DefaultPluginsConfig = Required<Omit<PluginsConfig, 'image'>> & {
+  // Image plugin should only have `validate` as a default prop, `upload` needs to be set by the user
+  image: Pick<ImageConfig, 'validate'>
+}
+const defaultPluginsConfig: DefaultPluginsConfig = {
   box: {
     allowedPlugins: [],
+  },
+  image: {
+    validate: validateFile,
   },
   multimedia: defaultMultimediaConfig,
   spoiler: {
