+++ packages/editor/src/plugins/image/index.ts
@@ -1,9 +1,9 @@
 import { editorPlugins } from '@editor/plugin/helpers/editor-plugins'
 import { EditorPluginType } from '@editor/types/editor-plugin-type'
 import { showToastNotice } from '@serlo/frontend/src/helper/show-toast-notice'
-import type { FileError } from '@serlo/frontend/src/serlo-editor-integration/image-with-serlo-config'
 
 import { ImageEditor } from './editor'
+import type { FileError } from './utils/validate-file'
 import {
   type EditorPlugin,
   type EditorPluginProps,
