+++ packages/editor/src/plugins/image/utils/validate-file.ts
@@ -1,0 +1,60 @@
+import { UploadValidator } from '@editor/plugin/upload'
+
+const maxFileSize = 2 * 1024 * 1024
+const allowedExtensions = ['gif', 'jpg', 'jpeg', 'png', 'svg', 'webp']
+
+enum FileErrorCode {
+  TOO_MANY_FILES,
+  NO_FILE_SELECTED,
+  BAD_EXTENSION,
+  FILE_TOO_BIG,
+  UPLOAD_FAILED,
+}
+
+export interface FileError {
+  errorCode: FileErrorCode
+  message: string
+}
+
+export const validateFile: UploadValidator<FileError[]> = (file) => {
+  let uploadErrors: FileErrorCode[] = []
+
+  if (!file) {
+    uploadErrors = [...uploadErrors, FileErrorCode.NO_FILE_SELECTED]
+  } else if (!matchesAllowedExtensions(file.name)) {
+    uploadErrors = [...uploadErrors, FileErrorCode.BAD_EXTENSION]
+  } else if (file.size > maxFileSize) {
+    uploadErrors = [...uploadErrors, FileErrorCode.FILE_TOO_BIG]
+  } else {
+    return { valid: true }
+  }
+
+  return { valid: false, errors: handleErrors(uploadErrors) }
+}
+
+function matchesAllowedExtensions(fileName: string) {
+  const extension = fileName.toLowerCase().slice(fileName.lastIndexOf('.') + 1)
+  return allowedExtensions.includes(extension)
+}
+
+function handleErrors(errors: FileErrorCode[]): FileError[] {
+  return errors.map((error) => ({
+    errorCode: error,
+    message: errorCodeToMessage(error),
+  }))
+}
+
+function errorCodeToMessage(error: FileErrorCode) {
+  switch (error) {
+    case FileErrorCode.TOO_MANY_FILES:
+      return 'You can only upload one file'
+    case FileErrorCode.NO_FILE_SELECTED:
+      return 'No file selected'
+    case FileErrorCode.BAD_EXTENSION:
+      return 'Not an accepted file type'
+    case FileErrorCode.FILE_TOO_BIG:
+      return 'Filesize is too big'
+    case FileErrorCode.UPLOAD_FAILED:
+      return 'Error while uploading'
+  }
+}
